+++ src/main/java/com/example/ilsan34/clientManager/service/PhotoService.java
@@ -2,17 +2,18 @@ package com.example.ilsan34.clientManager.service;
 
 import com.example.ilsan34.clientManager.model.Photo;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface PhotoService {
 
     void addNewPhoto(Photo photo);
 
-    Photo getById(Long id);
+    Photo getById(BigInteger id);
 
     void updatePhoto(Photo photo);
 
-    void deletePhoto(Long id);
+    void deletePhoto(BigInteger id);
 
     List<Photo> getAll();
 }
